DI Bridge: Changing to work in function of Dependencies instead of Shared logic
