+++ .gitignore
@@ -1,5 +1,6 @@
 .gradle/
 .idea/
-MooV.iml
+*.iml
+.DS_Store
 build/
 local.properties
