+++ app/src/main/java/br/com/moov/app/MooVApp.kt
@@ -17,14 +17,13 @@ package br.com.moov.app
 
 import android.app.Application
 import android.os.StrictMode
-import br.com.core.android.DependenciesProvider
 import br.com.core.android.logd
+import br.com.moov.dibridge.AppComponentDependenciesProvider
 
-class MooVApp : Application(), DependenciesProvider<Any> {
+class MooVApp : Application(), AppComponentDependenciesProvider {
 
     private val coreComponent by lazy {
-        DaggerAppComponent.factory()
-            .create(this)
+        DaggerAppComponent.factory().create(this)
     }
 
     override fun dependencies() = coreComponent
