--- core-android/src/main/java/br/com/core/android/DependencyExtensions.kt
@@ -1,44 +1,0 @@
-/*
- * Copyright 2020 Alex Almeida Tavella
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package br.com.core.android
-
-import android.content.Context
-import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.Fragment
-
-inline fun <reified D> AppCompatActivity.dependencies(): D {
-    return application.dependencies()
-}
-
-inline fun <reified D> Fragment.dependencies(): D {
-    return requireContext().applicationContext.dependencies()
-}
-
-inline fun <reified D> Context.dependencies(): D {
-    require(this is DependenciesProvider<*>) {
-        "$javaClass is not a subtype of ${DependenciesProvider::class.java}"
-    }
-    val deps = dependencies()
-    require(deps is D) {
-        "${deps?.javaClass} is not a subtype of ${D::class.java}. " +
-            "Did you forget to setup this module dependencies on the Application class?"
-    }
-    return deps
-}
-
-interface DependenciesProvider<out D> {
-    fun dependencies(): D
-}
