+++ feature/movie-details/src/main/java/br/com/moov/moviedetails/di/MovieDetailComponent.kt
@@ -15,13 +15,9 @@
  */
 package br.com.moov.moviedetails.di
 
-import br.com.moov.bookmark.movie.BookmarkMovieUseCase
-import br.com.moov.bookmark.movie.UnBookmarkMovieUseCase
-import br.com.moov.core.ImageUrlResolver
-import br.com.moov.moviedetails.navigation.MovieDetailsNavigator
+import br.com.moov.dibridge.MovieDetailsDependencies
 import br.com.moov.moviedetails.view.MovieDetailActivity
 import dagger.Component
-import retrofit2.Retrofit
 
 @Component(
     modules = [MovieDetailsInternalModule::class],
@@ -35,12 +31,4 @@ internal interface MovieDetailComponent {
     interface Factory {
         fun create(dependencies: MovieDetailsDependencies): MovieDetailComponent
     }
-}
-
-interface MovieDetailsDependencies {
-    fun retrofit(): Retrofit
-    fun imageUrlResolver(): ImageUrlResolver
-    fun bookmarkUseCase(): BookmarkMovieUseCase
-    fun unBookmarkUseCase(): UnBookmarkMovieUseCase
-    fun movieDetailNavigator(): MovieDetailsNavigator
 }
