+++ feature/movies/build.gradle.kts
@@ -7,6 +7,7 @@ plugins {
 dependencies {
     implementation(project(Modules.core))
     implementation(project(Modules.coreAndroid))
+    implementation(project(Modules.diBridge))
     implementation(project(Modules.FeatureApis.movies))
     implementation(project(Modules.FeatureApis.movieDetails))
     implementation(project(Modules.FeatureApis.bookmarkMovie))
