+++ feature/movies/src/main/java/br/com/moov/movies/di/MoviesComponent.kt
@@ -15,13 +15,9 @@
  */
 package br.com.moov.movies.di
 
-import br.com.moov.bookmark.movie.BookmarkMovieUseCase
-import br.com.moov.bookmark.movie.UnBookmarkMovieUseCase
-import br.com.moov.core.ImageUrlResolver
-import br.com.moov.moviedetails.navigation.MovieDetailsNavigator
+import br.com.moov.dibridge.MoviesDependencies
 import br.com.moov.movies.view.MoviesFragment
 import dagger.Component
-import retrofit2.Retrofit
 
 @Component(
     modules = [MoviesInternalModule::class],
@@ -35,12 +31,4 @@ internal interface MoviesComponent {
     interface Factory {
         fun create(dependencies: MoviesDependencies): MoviesComponent
     }
-}
-
-interface MoviesDependencies {
-    fun retrofit(): Retrofit
-    fun imageUrlResolver(): ImageUrlResolver
-    fun bookmarkUseCase(): BookmarkMovieUseCase
-    fun unBookmarkUseCase(): UnBookmarkMovieUseCase
-    fun movieDetailNavigator(): MovieDetailsNavigator
 }
